Remove convenience type aliases
Prior to this change, these aliases are sometimes used by users to
construct date/datetime objects, with the appearance that they are
constructing a timezone aware object. However, this is not the case,
since they are just aliases for the stdlib:

    >>> my_datetime = localtime.DateTime(2023, 6, 8, 11, 32)
    >>> my_datetime
    datetime.datetime(2023, 6, 8, 11, 32)
    >>> repr(my_datetime.tzinfo)
    'None'

This change removes these type aliases. This makes the module more
consistent (the stdlib imports were used for some type annotations) and
removes this footgun.