Enable more than two adjacent FiniteDateRanges to be unioned
Prior to this change, FiniteDateRange was able to correctly union with
an adjacent FiniteDateRange, however, the union produced was of the
base range type and was then unable to union with a successive adjacent
range.

This change ensures that unioning two FiniteDateRanges produces a
FiniteDateRange so that it can be unioned with a successive adjacent
ranges.