Include py.typed files in package data
Setup tools doesn't automatically include py.typed files in the package
data until the next release after Nov 19, 2023, which means that types
that were previously exported now present as Any.

This change adds py.typed files to each of the sub-packages, and includes
them in the package data for each.