Prevent querying range fields with tuples of (start, end)
The Django implementation of range fields allows querying them with
tuples of (start, end), presumably because the underlying pg_extras.Range
types aren't really supposed to be exposed to the user.

Considering we're attempting to enforce strict usages of types, we
explicitly prevent non-range types from being used in queries.