Alter classes derived from Range not to have instance dictionaries
Prior to this change, `Range` instances (because it declares
`__slots__`) did not have instance dict but classes derived
from `Range` did have instance dicts - a confusing mix of paradigms.

This change adds `__slots__ = ()` to classes that have `Range` as
a base class, and adds two tests to ensure that instances do not
have a `__dict__` attribute. It also adds an initialiser to
`HalfFiniteRange` to set `boundary` using the slot, rather than
trying to use an instance dict.