Split xocto.types into a separate package
Prior to this change, all types were in a single file, including the
Django specific types that depend on Django itself.

This change creates a new types sub-package and splits it into separate
modules. All of the types that depend on Django are now in a separate
module.

Now xocto.ranges and xocto.numbers can be imported without having to
install and configure Django.