Implement copy and deepcopy for ranges
Now that ranges are immutable, attempting to copy or deepcopy a range
(or something containing a range) will raise an error, since the default
implementation of copy is to set attributes on the new object.

This change implements copy and deepcopy by returning self, rather than
a new instance, because it is already immutable, which is consistent
with python's implementation of copy for things like functions.