Add LtreeField model field
This model field provides the basic toolkit for representing tree-like
data-structures within a PostgreSQL database, using the `ltree`
extension. It is not intended to be a complete framework for providing
this functionality, but instead a simple way to bring the `ltree` data
type into a Django model.

More on this extension can be found in the Postgres documentation:

https://www.postgresql.org/docs/current/ltree.html

The approach takes inspiration from the that outlined in

https://github.com/peopledoc/django-ltree-demo.

One difference is that the <@ and @> operators are reversed compared to
the demo, since this makes for more natural querying e.g.

    filter(path__ancestor=my_path)

returns the *descendants* of my_path (the demo uses path_descendant for
this).