Demonstrate issue with FiniteDateTimeRangeField
This commit shows the existence of an edge case introduced by this
change: 7558f7105dc4053bca190e927df0206923ea30a9

If we store the last hour of the DST period we won't be able to convert
it to a proper `FiniteDatetimeRange`. Because we set clocks back by one
hour then start == end.