Fix the daylight-saving bug on Ranges
Before this commit, it was not possible to create a range of 1 hour when the
daylight savings happened. This is because for Python these two objects are
equal:
 a = datetime.datetime(2024, 10, 27, 2, 59, fold=1, tzinfo=zoneinfo.ZoneInfo(key='Europe/Amsterdam')
 b = datetime.datetime(2024, 10, 27, 2, 59, tzinfo=zoneinfo.ZoneInfo(key='Europe/Amsterdam')
Even if converted to UTC they're different.

With this commit, we are converting the datetimes boundaries to UTC, so they
can be compared without issues.