Update date_of_day_before docstring
Prior to this change, the docstring was referring to an upper bound inclusive
convention for date ranges. Since there is no such convention in the
Kraken coding conventions, this can be confusing for the developers.

This change updates the docstring to reflect the fact that it is a thing
we usually do, but there's no written convention for it.