Add tests for type of FiniteDatetimeRange union
Ranges are equal (__eq__) based only on the fields
start, end and boundaries. It's possible for two ranges
to be equal without their types matching. So the existing
equality checks in these tests are not sufficient on their own.
In addition to checking for equality with an expected range, we should
check the type.