Reverse order of min in FiniteDatetimeRange.intersection
In https://github.com/octoenergy/xocto/pull/187 the implementation of
FiniteDatetimeRange.intersection was optimised. This introduced a very
subtle change in behaviour. Previously when left.end was equal to
right.end then the right.end was favoured. In
https://github.com/octoenergy/xocto/pull/187 the implementation was
(unintentionally) changed to favour left.end.

This change reverses that, to again favour right.end.

This really shouldn't matter, but it can be signidicant when
the TZ of left and right are different. Eventually we'd like
to prevent this kind of mixed TZ behaviour (see
https://github.com/octoenergy/xocto/issues/192), but in the short
term it seems reasonable to make this tiny change to make things
consistent again with how they were before.