Remove dependency from ranges module on localtime module
Prior to this change, the ranges module had a dependency on the
localtime module, which was only used for a single function call to
localtime.as_localtime that wraps django.utils.timezone.localtime.

This change removes the dependency. This should make it possible to
import xocto.ranges locally without Django running (which will speed up
some local testing).